Add Wagtail 2.12 compatibility
On Wagtail 2.12, switch to using raw_data in place of the deprecated stream_data. Switching to indexing the stream directly as a list (as per https://docs.wagtail.io/en/stable/releases/2.12.html#stream-data-on-streamfield-values-is-deprecated) is not really feasible here, as this would change public APIs that expect the JSON representation (e.g. https://wagtailstreamforms.readthedocs.io/en/latest/fields.html)