Fix handling of '&' URL separators in tests
Current Python versions as of February 2021 no longer treat the ; character as a URL parameter separator: https://python-security.readthedocs.io/vuln/urllib-query-string-semicolon-separator.html

This breaks the tests for the url_replace tag, as they attempt to parse a URL generated by a Django template. HTML autoescaping is enabled, so & characters in the URL will be escaped as &amp;. This discrepancy between escaped and unescaped HTML was previously masked because the closing ; was also treated as a separator.